# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fire Prevention API",
        "description": "API documentation with JWT authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get all devices",
                "description": "Returns a list of all registered devices",
                "operationId": "d3978c2fbd1796da59a3cf3576e55df4",
                "responses": {
                    "200": {
                        "description": "List of devices retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DeviceResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No devices found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No devices found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Create a new device",
                "description": "Creates a new device with the provided information and generates a unique token",
                "operationId": "4f947fbe016f1c35e9a6ad86cc74885a",
                "requestBody": {
                    "description": "Device creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the device (min: 10, max: 100 characters)",
                                        "type": "string",
                                        "example": "Sensor Temperatura Cocina"
                                    },
                                    "location": {
                                        "description": "Location of the device (min: 10, max: 100 characters)",
                                        "type": "string",
                                        "example": "Cocina Principal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The name must be at least 10 characters"
                                            ]
                                        },
                                        "location": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The location must be at least 10 characters"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get a single device",
                "description": "Returns detailed information about a specific device",
                "operationId": "17ad13acb1600bf31cd958047e432d22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the device to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update a device",
                "description": "Updates an existing device with the provided information",
                "operationId": "e1fb9e989e7a45c0c9c826c63f32e2f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the device to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Device update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sensor Temperatura Sala"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Sala Principal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Device updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete a device",
                "description": "Deletes an existing device",
                "operationId": "12ef76711eba12287d34f2f016a91877",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the device to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/activate/{token}": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Activate a device",
                "description": "Activates a device using its unique token",
                "operationId": "13229d2ea781e7d37b9bd20c6061387b",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token of the device to activate",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device activated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device activated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/devices/deactivate/{token}": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Deactivate a device",
                "description": "Deactivates a device using its unique token",
                "operationId": "d3592a42411be1721d20eb738ee5e836",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token of the device to deactivate",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device deactivated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device deactivated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/devices/status/{token}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get device status by token",
                "description": "Returns the status of a device using its unique token",
                "operationId": "e0d6e24871c1da463744f51ddb85b2d9",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token of the device to check",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "Activo",
                                                "Inactivo"
                                            ],
                                            "example": "Activo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Device not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeviceResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sensor Temperatura Cocina"
                    },
                    "location": {
                        "type": "string",
                        "example": "Cocina Principal"
                    },
                    "token": {
                        "type": "string",
                        "example": "a1b2c3d4e5f6g7h8i9j0"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Activo",
                            "Inactivo"
                        ],
                        "example": "Activo"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-29T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-29T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}